Moved Asm tests to a directory
