+++ test/unit/Asm/AsmGroupsTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace SendGridPhp\Tests;
+namespace SendGridPhp\Tests\Asm;
+
+use SendGridPhp\Tests\BaseTestClass;
 
 class AsmGroupsTest extends BaseTestClass
 {
